refactor: 예약 step4 activity fragment로 변경
