+++ app/src/main/res/layout/activity_reservation_step5.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.reservation.ReservationStep4Activity">
+    tools:context=".ui.reservation.ReservationStep4Fragment">
 
     <ImageView
         android:id="@+id/back_btn"
